Content Added
Added three poets to the index and tried to make another page, but only succeeded in making another button, but it leads to an error.

The products.cs in Models is someting I did because I was following along with the asp.net 101 from Microsoft, but it's not something I can understand yet. I'm just doing the html first and then we'll worry about doing actual coder stuff later.